restructuring and fixing authorization for adding, editing and deleting thoughts
