+++ src/sections/NavBar.jsx
@@ -1,3 +1,4 @@
+import { Link } from "react-router-dom"
 import { useAuth } from "../context/AuthContext"
 import NavLink from "../components/NavLink"
 
@@ -6,19 +7,22 @@ const NavBar = () => {
 
   return (
     <header className="sticky top-0 w-full left-0 z-2">
-      <nav className="py-2 px-5 bg-(--color-background) text-(--color-text)">
-        <ul className="flex justify-end gap-6">
+      <nav className="flex justify-between py-2 px-2 sm:px-5 bg-(--color-background) text-(--color-text)">
+        <Link to="/" >
+          <h1 className="font-semibold" >Happy Thoughts</h1>
+        </Link>
+        <ul className="flex justify-end gap-3 sm:gap-6">
           {isLoggedIn ? (
             <>
-              <NavLink text="Thoughts" linkTo="/thoughts" />
+              <NavLink text="My thoughts" linkTo="/thoughts" />
               <li>
                 <button className="cursor-pointer" type="button" onClick={logout}>Log out</button>
               </li>
             </>
           ) : (
             <>
-              <NavLink text="Log in" linkTo="/login" />
               <NavLink text="Sign up" linkTo="/register" />
+              <NavLink text="Log in" linkTo="/login" />
             </>
           )}
         </ul>
